+++ src/indra_cogex/apps/__init__.py
@@ -1,8 +1,3 @@
 # -*- coding: utf-8 -*-
 
 """Apps built on top of INDRA CoGEx."""
-from pathlib import Path
-
-APPS_DIR = Path(__file__).parent.absolute()
-TEMPLATES_DIR = APPS_DIR / "templates"
-STATIC_DIR = APPS_DIR / "static"
