+++ src/indra_cogex/apps/data_display/cli.py
@@ -1,10 +1,10 @@
 from flask import Flask
 from flask_bootstrap import Bootstrap4
+from indralab_auth_tools.auth import auth, config_auth
 from more_click import make_web_command
 
-from indra_cogex.apps import STATIC_DIR, TEMPLATES_DIR
+from indra_cogex.apps.constants import STATIC_DIR, TEMPLATES_DIR
 from indra_cogex.apps.data_display import data_display_blueprint
-from indralab_auth_tools.auth import auth, config_auth
 
 app = Flask(__name__, template_folder=TEMPLATES_DIR, static_folder=STATIC_DIR)
 bootstrap = Bootstrap4(app)
