--- src/indra_cogex/apps/gla/constants.py
@@ -1,12 +1,0 @@
-"""Web constants."""
-
-import os
-
-__all__ = [
-    "INDRA_COGEX_WEB_LOCAL",
-]
-
-INDRA_COGEX_WEB_LOCAL = os.environ.get("INDRA_COGEX_WEB_LOCAL", "").lower() in {
-    "t",
-    "true",
-}
