+++ src/indra_cogex/apps/landing_page/cli.py
@@ -3,7 +3,7 @@ from flask_bootstrap import Bootstrap4
 from indralab_auth_tools.auth import auth, config_auth
 from more_click import make_web_command
 
-from indra_cogex.apps import STATIC_DIR, TEMPLATES_DIR
+from indra_cogex.apps.constants import STATIC_DIR, TEMPLATES_DIR
 from indra_cogex.apps.landing_page import landing_blueprint
 from indra_cogex.apps.proxies import INDRA_COGEX_EXTENSION
 from indra_cogex.client import Neo4jClient
